requirements.txt
streamlit
google-generativeai
streamlit-webrtc
numpy
OpenAI-Whisper
soundfile
pydub
av
aiortc
python-dotenv